refactor: convert hooks to actions
